requirements.txt
flask
flask-smorest
python-dotenv
sqlalchemy
flask-sqlalchemy
gunicorn
psycopg2
Werkzeug